+++ packages/sdk/src/sdk-blockchains/ethereum/balance.ts
@@ -1,35 +1,53 @@
 import type { RaribleSdk } from "@rarible/protocol-ethereum-sdk"
 import type { UnionAddress } from "@rarible/types"
 import type { AssetType } from "@rarible/api-client"
-import type { Erc20AssetType, EthAssetType } from "@rarible/ethereum-api-client"
 import type { BigNumberValue } from "@rarible/utils"
-import { convertToEthereumAddress } from "./common"
+import type { IBlockchainTransaction } from "@rarible/sdk-transaction"
+import { BlockchainEthereumTransaction } from "@rarible/sdk-transaction"
+import type { EthereumNetwork } from "@rarible/protocol-ethereum-sdk/build/types"
+import type {
+	AssetType as EthereumAssetType,
+} from "@rarible/ethereum-api-client"
+import type { ConvertRequest } from "../../types/balances"
+import { convertToEthereumAddress, convertToEthereumAssetType } from "./common"
 
 export class EthereumBalance {
-	constructor(private sdk: RaribleSdk) {
+	constructor(
+		private sdk: RaribleSdk,
+		private network: EthereumNetwork,
+	) {
 		this.getBalance = this.getBalance.bind(this)
-	}
-
-	convertAssetType(assetType: AssetType): EthAssetType | Erc20AssetType {
-		switch (assetType["@type"]) {
-			case "ETH": {
-				return { assetClass: "ETH" }
-			}
-			case "ERC20": {
-				return {
-					assetClass: "ERC20",
-					contract: convertToEthereumAddress(assetType.contract),
-				}
-			}
-			default: {
-				throw new Error(`Unsupported asset type=${assetType["@type"]}`)
-			}
-		}
+		this.convert = this.convert.bind(this)
 	}
 
 	async getBalance(address: UnionAddress, assetType: AssetType): Promise<BigNumberValue> {
+		const convertedAssetType = convertToEthereumAssetType(assetType)
+		if (convertedAssetType.assetClass !== "ETH" && convertedAssetType.assetClass !== "ERC20") {
+			throw new Error("Unsupported asset type for getting balance")
+		}
 		const ethAddress = convertToEthereumAddress(address)
-		const convertedAssetType = this.convertAssetType(assetType)
 		return this.sdk.balances.getBalance(ethAddress, convertedAssetType)
+	}
+
+	async convert(request: ConvertRequest): Promise<IBlockchainTransaction> {
+		const wethContract = this.sdk.balances.getWethContractAddress()
+		let from: EthereumAssetType
+		let to: EthereumAssetType
+
+		if (request.isWrap) {
+			from = { assetClass: "ETH" }
+			to = {
+				assetClass: "ERC20",
+				contract: wethContract,
+			}
+		} else {
+			from = {
+				assetClass: "ERC20",
+				contract: wethContract,
+			}
+			to = { assetClass: "ETH" }
+		}
+		const tx = await this.sdk.balances.convert(from, to, request.value)
+		return new BlockchainEthereumTransaction(tx, this.network)
 	}
 }