+++ packages/wallet/src/index.ts
@@ -1,15 +1,13 @@
 import type { Ethereum } from "@rarible/ethereum-provider"
 import type { Fcl } from "@rarible/fcl-types"
-import { Blockchain } from "@rarible/api-client"
+import { BlockchainGroup } from "@rarible/api-client"
 import type { TezosProvider } from "@rarible/tezos-sdk"
 import type { AbstractWallet, UserSignature } from "./domain"
 
-export class EthereumWallet<K extends Blockchain.ETHEREUM | Blockchain.POLYGON,
-	T extends Ethereum = Ethereum> implements AbstractWallet {
-	constructor(
-		public readonly ethereum: T,
-		public readonly blockchain: K,
-	) {
+export class EthereumWallet<T extends Ethereum = Ethereum> implements AbstractWallet {
+	readonly blockchain = BlockchainGroup.ETHEREUM
+
+	constructor(public readonly ethereum: T) {
 	}
 
 	async signPersonalMessage(message: string): Promise<UserSignature> {
@@ -25,7 +23,7 @@ export class EthereumWallet<K extends Blockchain.ETHEREUM | Blockchain.POLYGON,
 }
 
 export class FlowWallet implements AbstractWallet {
-	readonly blockchain = Blockchain.FLOW
+	readonly blockchain = BlockchainGroup.FLOW
 
 	constructor(public readonly fcl: Fcl) {
 	}
@@ -72,7 +70,7 @@ export interface TezosSignatureResult {
 }
 
 export class TezosWallet implements AbstractWallet {
-	readonly blockchain = Blockchain.TEZOS
+	readonly blockchain = BlockchainGroup.TEZOS
 
 	constructor(public readonly provider: TezosProvider) {
 	}
@@ -100,14 +98,12 @@ export class TezosWallet implements AbstractWallet {
 }
 
 export type BlockchainWallet =
-	EthereumWallet<Blockchain.ETHEREUM> |
-	EthereumWallet<Blockchain.POLYGON> |
+	EthereumWallet |
 	FlowWallet |
 	TezosWallet
 
 export type WalletByBlockchain = {
 	"FLOW": FlowWallet
-	"ETHEREUM": EthereumWallet<Blockchain.ETHEREUM>,
-	"POLYGON": EthereumWallet<Blockchain.POLYGON>,
+	"ETHEREUM": EthereumWallet,
 	"TEZOS": TezosWallet
 }